arb-bot-project/backend/requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.23.2
websockets==11.0.3
python-dotenv==1.0.0
numpy==1.25.2
pandas==2.1.0
xgboost==2.0.0
scikit-learn==1.3.1
httpx==0.25.0
python-binance==1.0.19
aiohttp==3.8.5
